requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs<19.4,>=18.2
audio-metadata==0.11.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
discord.py[voice]==1.5.0
distlib==0.3.0
filelock==3.0.12
idna==2.10
multidict==4.7.6
ordered-set==4.0.1
packaging==20.4
pycparser==2.20
PyNaCl==1.3.0
pyparsing==2.4.7
pyyaml==5.3.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
virtualenv==20.0.21
yapf==0.30.0
yarl==1.5.1
